little_endian: avoid signed shift overflow in getDW
(c3 << 8) << 16 shifts into the sign bit for bytes >= 0x80, which is
undefined behaviour (flagged by UBSan on every AIN/AFA load). Do the
assembly in unsigned arithmetic and convert once.

+++ src/little_endian.h
@@ -33,7 +33,7 @@ static inline int32_t LittleEndian_getDW(const uint8_t *b, int i)
 	c3 = *(b + i  + 3);
 	d0 = c0 + (c1 << 8);
 	d1 = c2 + (c3 << 8);
-	return d0 + (d1 << 16);
+	return (int32_t)((uint32_t)d0 + ((uint32_t)d1 << 16));
 }
 
 static inline int32_t LittleEndian_get3B(const uint8_t *b, int i)
